+++ src/onegov.election_day/onegov/election_day/formats/__init__.py
@@ -9,6 +9,7 @@ from onegov.election_day.formats.vote.default import import_vote_default
 from onegov.election_day.formats.vote.internal import import_vote_internal
 from onegov.election_day.formats.vote.wabsti import import_vote_wabsti
 from onegov.election_day.formats.vote.wabstic import import_vote_wabstic
+from onegov.election_day.formats.vote.wabstim import import_vote_wabstim
 
 
 __all__ = [
@@ -22,4 +23,5 @@ __all__ = [
     'import_vote_internal',
     'import_vote_wabsti',
     'import_vote_wabstic',
+    'import_vote_wabstim',
 ]