+++ src/onegov.election_day/onegov/election_day/formats/election/wabsti_majorz.py
@@ -128,6 +128,16 @@ def import_election_wabsti_majorz(
     election, entities, file, mimetype,
     elected_file=None, elected_mimetype=None
 ):
+    """ Tries to import the given csv, xls or xlsx file.
+
+    This is the format used by Wabsti for majorz elections. Since there is no
+    format description, importing these files is somewhat experimental.
+
+    :return:
+        A list containing errors.
+
+    """
+
     errors = []
     candidates = {}
     results = []