+++ src/onegov.election_day/onegov/election_day/formats/vote/default.py
@@ -25,6 +25,9 @@ def import_vote_default(vote, entities, ballot_type, file, mimetype):
     """ Tries to import the given csv, xls or xlsx file to the given ballot
     result type.
 
+    This is a custom format defined by us to easily create vote results by
+    hand.
+
     :return:
         A list containing errors.