+++ src/onegov.election_day/onegov/election_day/formats/vote/wabstim.py
@@ -1,7 +1,6 @@
 from onegov.ballot import Ballot
 from onegov.ballot import BallotResult
 from onegov.election_day import _
-from onegov.election_day.formats.common import EXPATS
 from onegov.election_day.formats.common import FileImportError
 from onegov.election_day.formats.common import load_csv
 from onegov.election_day.utils import clear_vote
@@ -9,39 +8,42 @@ from onegov.election_day.utils import guessed_group
 
 
 HEADERS = (
-    'vorlage-nr.',
-    'bfs-nr.',
+    'freigegeben',
+    'stileer',
+    'stiungueltig',
+    'stijahg',
+    'stineinhg',
+    'stiohneawhg',
+    'stijan1',
+    'stineinn1',
+    'stiohneawN1',
+    'stijan2',
+    'stineinn2',
+    'stiohneawN2',
     'stimmberechtigte',
-    'leere sz',
-    'ungultige sz',
-    'ja',
-    'nein',
-    'initoantw',
-    'gegenvja',
-    'gegenvnein',
-    'gegenvoantw',
-    'stichfrja',
-    'stichfrnein',
-    'stichfroantw',
-    'stimmbet',
+    'bfs',
 )
 
 
-def import_vote_wabsti(vote, entities, vote_number, complex, file, mimetype):
+def import_vote_wabstim(vote, entities, complex, file, mimetype):
     """ Tries to import the given csv, xls or xlsx file.
 
-    This is the format used by Wabsti. Since there is no format description,
-    importing these files is somewhat experimental.
+    This is the format used by Wabsti for municipalities. Since there is no
+    format description, importing these files is somewhat experimental.
 
     :return:
         A list containing errors.
 
     """
-    csv, error = load_csv(file, mimetype, expected_headers=HEADERS)
+    csv, error = load_csv(
+        file, mimetype, expected_headers=HEADERS,
+        rename_duplicate_column_names=True
+    )
     if error:
         # Wabsti files are sometimes UTF-16
         csv, utf16_error = load_csv(
-            file, mimetype, expected_headers=HEADERS, encoding='utf-16-le'
+            file, mimetype, expected_headers=HEADERS, encoding='utf-16-le',
+            rename_duplicate_column_names=True
         )
         if utf16_error:
             return [error]
@@ -51,50 +53,27 @@ def import_vote_wabsti(vote, entities, vote_number, complex, file, mimetype):
         used_ballot_types.extend(['counter-proposal', 'tie-breaker'])
 
     ballot_results = {key: [] for key in used_ballot_types}
-    errors = []
     added_entity_ids = set()
+    errors = []
     skipped = 0
 
     for line in csv.lines:
-
         line_errors = []
-
-        # Skip the results of other votes
-        try:
-            number = int(line.vorlage_nr_ or 0)
-        except ValueError:
-            line_errors.append(_("Invalid values"))
-        else:
-            if number != vote_number:
-                continue
-
-        # Skip not yet counted results
-        try:
-            turnout = float(line.stimmbet or 0)
-        except ValueError:
-            line_errors.append(_("Invalid values"))
-        else:
-            if not turnout:
-                skipped += 1
-                continue
 
         # the id of the entity
         entity_id = None
         try:
-            entity_id = int(line.bfs_nr_ or 0)
+            entity_id = int(line.bfs or 0)
         except ValueError:
             line_errors.append(_("Invalid id"))
         else:
-            if entity_id not in entities and entity_id in EXPATS:
-                entity_id = 0
-
             if entity_id in added_entity_ids:
                 line_errors.append(
                     _("${name} was found twice", mapping={
                         'name': entity_id
                     }))
 
-            if entity_id and entity_id not in entities:
+            if entity_id not in entities:
                 line_errors.append(
                     _("${name} is unknown", mapping={
                         'name': entity_id
@@ -102,21 +81,26 @@ def import_vote_wabsti(vote, entities, vote_number, complex, file, mimetype):
             else:
                 added_entity_ids.add(entity_id)
 
+        try:
+            counted = True if line.freigegeben else False
+        except ValueError:
+            line_errors.append(_("Invalid values"))
+
         # the yeas
         yeas = {}
         try:
-            yeas['proposal'] = int(line.ja or 0)
-            yeas['counter-proposal'] = int(line.gegenvja or 0)
-            yeas['tie-breaker'] = int(line.stichfrja or 0)
+            yeas['proposal'] = int(line.stijahg or 0)
+            yeas['counter-proposal'] = int(line.stijan1 or 0)
+            yeas['tie-breaker'] = int(line.stijan2 or 0)
         except ValueError:
             line_errors.append(_("Could not read yeas"))
 
         # the nays
         nays = {}
         try:
-            nays['proposal'] = int(line.nein or 0)
-            nays['counter-proposal'] = int(line.gegenvnein or 0)
-            nays['tie-breaker'] = int(line.stichfrnein or 0)
+            nays['proposal'] = int(line.stineinhg or 0)
+            nays['counter-proposal'] = int(line.stineinn1 or 0)
+            nays['tie-breaker'] = int(line.stineinn2 or 0)
         except ValueError:
             line_errors.append(_("Could not read nays"))
 
@@ -125,24 +109,20 @@ def import_vote_wabsti(vote, entities, vote_number, complex, file, mimetype):
             elegible_voters = int(line.stimmberechtigte or 0)
         except ValueError:
             line_errors.append(_("Could not read the elegible voters"))
-        else:
-            # Ignore the expats if no eligible voters
-            if not entity_id and not elegible_voters:
-                continue
 
         # the empty votes
         empty = {}
         try:
-            e_ballots = int(line.leere_sz or 0)
-            empty['proposal'] = int(line.initoantw or 0) + e_ballots
-            empty['counter-proposal'] = int(line.gegenvoantw or 0) + e_ballots
-            empty['tie-breaker'] = int(line.stichfroantw or 0) + e_ballots
+            e_ballots = int(line.stileer or 0)
+            empty['proposal'] = int(line.stiohneawhg or 0) + e_ballots
+            empty['counter-proposal'] = int(line.stiohneawn1 or 0) + e_ballots
+            empty['tie-breaker'] = int(line.stiohneawn2 or 0) + e_ballots
         except ValueError:
             line_errors.append(_("Could not read the empty votes"))
 
         # the invalid votes
         try:
-            invalid = int(line.ungultige_sz or 0)
+            invalid = int(line.stiungueltig or 0)
         except ValueError:
             line_errors.append(_("Could not read the invalid votes"))
 
@@ -160,7 +140,7 @@ def import_vote_wabsti(vote, entities, vote_number, complex, file, mimetype):
                 ballot_results[ballot_type].append(
                     BallotResult(
                         group=entities.get(entity_id, {}).get('name', ''),
-                        counted=True,
+                        counted=counted,
                         yeas=yeas[ballot_type],
                         nays=nays[ballot_type],
                         elegible_voters=elegible_voters,