+++ src/onegov.election_day/onegov/election_day/forms/upload.py
@@ -366,6 +366,7 @@ class UploadVoteForm(Form):
             self.file_format.choices = [
                 ('default', _("Default")),
                 ('internal', "OneGov Cloud"),
+                ('wabsti_m', "Wabsti"),
             ]
         else:
             self.file_format.choices = [